+++ lib/shared/widgets/brand_logo.dart
@@ -16,7 +16,7 @@ class BrandLogo extends StatelessWidget {
         AppBranding.logoAsset,
         width: size,
         height: size,
-        fit: BoxFit.cover,
+        fit: BoxFit.contain,
         filterQuality: FilterQuality.high,
       ),
     );